Moved service registration to core from data layer.  Also refactored out httpclient and identity configuration to service extension class inside MVC layer to make program clean
