--- MyShow.Data/Extensions/ServicesExtensions.cs
@@ -1,16 +1,0 @@
-﻿using Microsoft.Extensions.Configuration;
-using Microsoft.Extensions.DependencyInjection;
-using NetCore.AutoRegisterDi;
-
-namespace MyShow.Data.Extensions;
-public static class ServicesExtensions
-{
-    public static void AddDataServices(this IServiceCollection services, IConfiguration config)
-    {
-        services.RegisterAssemblyPublicNonGenericClasses()
-            .Where(c => c.Name.EndsWith("Repository") || c.Name.EndsWith("Action") || c.Name.EndsWith("Queries"))
-            .AsPublicImplementedInterfaces(ServiceLifetime.Scoped);
-
-        services.AddDbContext<AppDbContext>(options => options.UseSqlServer(config.GetConnectionString("DefaultConnection")));
-    }
-}
